related-patent-discovery/requirements-py310.txt
fastapi==0.100.0
uvicorn==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.4
PyMuPDF==1.22.3
pdf2image==1.16.3
pytesseract==0.3.10
python-dotenv==1.0.0
web3==5.31.4
requests==2.31.0
huggingface-hub==0.14.1